+++ DotNetDo/Cli/RootCommand.cs
@@ -13,6 +13,7 @@ static class RootCommand
             {
                 ":init" => InitCommand.Run(args),
                 ":new" => NewCommand.Run(args),
+                ":install" => await InstallCommand.Run(args),
                 ":help" => HelpCommand.Run(args),
                 var command when command.StartsWith(':') => Fail($"Unknown command '{command}'."),
                 var taskName => await RunCommand.RunTask(taskName, args[1..])